Remove dependency on parent for AttachmentsForPatient.
